Crate lut

lut (lookup tables for every one)

Currently this provides a complex macro. In the future this should be a proc-macro which will enable most of the benefits behind this crate with nive ergonomics. Also currently this crate contains some feature-gated lookup tables. They will be moved out of this crate in the future.

Modules

media_type_chars
a lookup table for parsing/validating Media Types (also know as MIME-Types)

Macros

accessor_all

accessor_any

merge_tables
Example

new_table
Example

Structs

All

Any

EmptyFlag

NoFlagsSet

Traits

Access

ConstFlagCount

Flag

Table

TableValue