Module lut::media_type_chars

a lookup table for parsing/validating Media Types (also know as MIME-Types)

To use lookup the table compile the crate with the media-type-chars feature enabled, else wise this module will be empty.

Structs

CText
CText

DQuoteOrEscape
The characters mainly needing escaping i.e. '"' and '\'

HttpToken
Rfc7230Token

MediaTypeChars

ObsNoWsCtl
ObsNoWsCtl

ObsQText

ObsQTextWs

QText
QText

QTextWs

RestrictedToken
Restricted Token

Token
Token (Mime)

VChar

VCharWs

Ws
Ws (\t and \r)